requirements.txt
django == 2.2.3
mysqlclient
pillow
django-pure-pagination